Create PlansUpdate.vue.

+++ src/pages/Plans/PlansUpdate.vue
@@ -1,40 +1,56 @@
 <template>
   <div class="plans-update">
-    <div class="plans-update__header">
-      <h1 class="plans-update__header-heading">{{plan.name}}</h1>
-      <div class="plans-update__header-button">
-        <button-unit
-            class="plans-update__header-button-delete"
-            color="red"
-            @click="handleDelete"
-        >
-          Delete
-        </button-unit>
-        <button-unit
-            class="plans-update__header-button-edit"
-            color="purple"
-            @click="handleEdit"
-        >
-          Edit
-        </button-unit>
+    <Form
+        class="plans-update__form"
+        @submit="onSubmit"
+        :validation-schema="schema"
+    >
+      <div class="plans-update__header">
+        <h1 class="plans-update__heading">{{plan.name}}</h1>
+        <div class="plans-update__button">
+          <button-unit
+              class="plans-update__button-delete"
+              color="red"
+              @click="handleDelete"
+          >
+            Delete
+          </button-unit>
+          <button-unit
+              class="plans-update__button-edit"
+              color="purple"
+              type="submit"
+          >
+            Edit
+          </button-unit>
+        </div>
       </div>
-    </div>
-    <form class="plans-update__form">
       <label
-          class="plans-update__form-label"
+          class="plans-update__label"
           for="name"
       >
         Name
       </label>
-      <input
-          class="plans-update__form-input"
+      <Field
+          class="plans-update__input"
           type="text"
           id="name"
           name="name"
-          placeholder="name"
-          v-model="plan.name"
       />
-    </form>
+      <ErrorMessage name="name" />
+      <label
+          class="plans-update__label"
+          for="value"
+      >
+        Value
+      </label>
+      <Field
+          class="plans-update__input"
+          type="text"
+          id="value"
+          name="value"
+      />
+      <ErrorMessage name="value" />
+    </Form>
   </div>
 </template>
 
@@ -42,12 +58,16 @@
 import {defineComponent, ref} from "vue";
 import axios from "axios";
 import ButtonUnit from "@/components/atoms/ButtonUnit/ButtonUnit.vue";
+import { Field, Form, ErrorMessage } from 'vee-validate';
+import * as yup from 'yup';
 
 export default  defineComponent({
   name: 'PlansUpdate',
   components: {
-    ButtonUnit
-
+    ButtonUnit,
+    Field,
+    Form,
+    ErrorMessage
   },
   setup() {
     const plan = ref({id: null, name: '', value: 100});
@@ -64,19 +84,32 @@ export default  defineComponent({
       plan,
     }
   },
+  data() {
+    const schema = yup.object({
+      name: yup.string().required().label('Name'),
+      value: yup.string().required().label('Value'),
+    });
+    return {
+      schema,
+    };
+  },
   methods: {
-    handleEdit(): void {
-      axios.post(`${process.env.VUE_APP_BACKEND_URL}/plans/${this.plan}`, this.plan);
+    onSubmit(values: any): void {
+      axios.post(`${process.env.VUE_APP_BACKEND_URL}/plans/${this.plan}`, values);
       this.$router.push({ name: 'PlansRetrieveAll' });
     },
     handleDelete(): void {
-      axios.delete(`${process.env.VUE_APP_BACKEND_URL}/plans/${this.plan}`);
-      this.$router.push({ name: 'PlansRetrieveAll' });
+      if (confirm("Do you want to delete?")) {
+        axios.delete(`${process.env.VUE_APP_BACKEND_URL}/plans/${this.plan}`);
+        this.$router.push({ name: 'PlansRetrieveAll' });
+      }
     },
   },
 })
 </script>
 
 <style lang="scss">
-.plans-update {}
+.plans-update {
+  @include page-update();
+}
 </style>
